Move common calculator unit tests to sindex.
